requirements.txt
streamlit
pandas
scikit-learn
joblib
requests
beautifulsoup4
google-generativeai